Add: functionality to change the reserved state of the mission Remove: utils directory Refactor redux mission and rockets filenames for conflicts reasons
