--- src/utils/generator.js
@@ -1,6 +1,0 @@
-const reservedStatus = [true, false];
-
-const generateRandomReversedStatus = () => reservedStatus[
-  Math.floor(Math.random() * reservedStatus.length)];
-
-export default generateRandomReversedStatus;
