--- src/utils/usePathName.js
@@ -1,8 +1,0 @@
-import { useLocation } from 'react-router-dom';
-
-const usePathname = () => {
-  const location = useLocation();
-  return { pathname: location.pathname };
-};
-
-export default usePathname;
